// interfaces/response.ts
export interface TotalVolume {
  aed: any;
  ars: any;
  aud: any;
  bch: number;
  bdt: any;
  bhd: any;
  bmd: any;
  bnb: number;
  brl: any;
  btc: number;
  cad: any;
  chf: any;
  clp: any;
  cny: any;
  czk: any;
  dkk: any;
  dot: number;
  eos: any;
  eth: number;
  eur: any;
  gbp: any;
  hkd: any;
  huf: any;
  idr: any;
  ils: any;
  inr: any;
  jpy: any;
  krw: any;
  kwd: any;
  lkr: any;
  ltc: number;
  mmk: any;
  mxn: any;
  myr: any;
  ngn: any;
  nok: any;
  nzd: any;
  php: any;
  pkr: any;
  pln: any;
  rub: any;
  sar: any;
  sek: any;
  sgd: any;
  thb: any;
  try: any;
  twd: any;
  uah: any;
  usd: any;
  vef: any;
  vnd: any;
  xag: number;
  xau: number;
  xdr: any;
  xlm: any;
  xrp: any;
  yfi: number;
  zar: any;
  bits: any;
  link: number;
  sats: any;
}
// ...
export interface TotalMarketCap {
  btc: number;
  eth: number;
  ltc: number;
  bch: number;
  bnb: number;
  eos: number;
  xrp: number;
  xlm: number;
  link: number;
  dot: number;
  yfi: number;
  usd: number;
  aed: number;
  ars: number;
  aud: number;
  bdt: number;
  bhd: number;
  bmd: number;
  brl: number;
  cad: number;
  chf: number;
  clp: number;
  cny: number;
  czk: number;
  dkk: number;
  eur: number;
  gbp: number;
  hkd: number;
  huf: number;
  idr: number;
  ils: number;
  inr: number;
  jpy: number;
  krw: number;
  kwd: number;
  lkr: number;
  mmk: number;
  mxn: number;
  myr: number;
  ngn: number;
  nok: number;
  nzd: number;
  php: number;
  pkr: number;
  pln: number;
  rub: number;
  sar: number;
  sek: number;
  sgd: number;
  thb: number;
  try: number;
  twd: number;
  uah: number;
  vef: number;
  vnd: number;
  zar: number;
  xdr: number;
  xag: number;
  xau: number;
  bits: number;
  sats: number;
}

export interface MarketCapPercentage {
  btc: number;
  eth: number;
  bnb: number;
  usdt: number;
  sol: number;
  ada: number;
  usdc: number;
  xrp: number;
  luna: number;
  dot: number;
}

export interface GlobalMarketData {
  active_cryptocurrencies: number;
  upcoming_icos: number;
  ongoing_icos: number;
  ended_icos: number;
  markets: number;
  total_market_cap: TotalMarketCap;
  total_volume: TotalVolume;
  market_cap_percentage: MarketCapPercentage;
  market_cap_change_percentage_24h_usd: number;
  // updated_at: number;
}